Added ability to export parent SKUs in dataflow
Added ability to export with parent SKUs for configurable products. To
use this, add the following line in the dataflow modifier (as an
example, reconfigure the CSV field name to your requirements):
<var name="add_parent_sku">parent_sku</var>

Hopefully this will save people some pain when trying to batch export
including the parent SKUs for the likes of Magmi.